Moved UUID generation logic to secret package
